Unify the activity and metrics server stubs into one ServerStub
