+++ Tests/ScoutTests/UI/Activity/ActivityProviderTests.swift
@@ -14,7 +14,7 @@ import Testing
 struct ActivityProviderTests {
     @Test("Returns the native series from a Scout server unchanged")
     func fetchUsesServerSeries() async throws {
-        let database = ServerStub(series: [
+        let database = ServerStub(activity: [
             ActivityPoint(date: ms(2026, 6, 10), dau: 2, wau: 2, mau: 2),
             ActivityPoint(date: ms(2026, 6, 11), dau: 1, wau: 3, mau: 2),
             ActivityPoint(date: ms(2026, 7, 1), dau: 0, wau: 0, mau: 5),
@@ -55,36 +55,5 @@ struct ActivityProviderTests {
 
     private func ms(_ year: Int, _ month: Int, _ day: Int) -> Int64 {
         Int64((date(year, month, day).timeIntervalSince1970 * 1000).rounded())
-    }
-}
-
-/// A `DatabaseReader` that returns a native active-user series, standing in for a
-/// Scout server backend.
-///
-private final class ServerStub: DatabaseReader, @unchecked Sendable {
-    let series: [ActivityPoint]
-
-    init(series: [ActivityPoint]) {
-        self.series = series
-    }
-
-    func activity(in range: Range<Date>) async throws -> [ActivityPoint] {
-        series
-    }
-
-    func metricSeries<T: SeriesScalar>(_ valueType: T.Type, category: String, in range: Range<Date>) async throws -> [MetricSeries] {
-        []
-    }
-
-    func lookup(recordName: String, fields: [String]?) async throws -> Record {
-        throw RecordNotFoundError()
-    }
-
-    func read(matching query: RecordQuery, fields: [String]?) async throws -> RecordChunk {
-        RecordChunk(records: [], cursor: nil)
-    }
-
-    func readMore(from cursor: RecordCursor, fields: [String]?) async throws -> RecordChunk {
-        RecordChunk(records: [], cursor: nil)
     }
 }
